Almost finished now!
Just added collisions with cars, riding along the logs, and map boundaries. The only thing that's left now is making the water a hazard. I'm not sure how much memory all the for loops will take on my server but still I've had a lot of fun with this so far haha